services/campaign-runner/pyproject.toml
[project]
name = "campaign-runner"
version = "0.1.0"
description = "Worker that polls email_queue table and sends emails via SparkPost API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "httpx>=0.27.0",
    "google-genai>=1.55.0",
]

[tool.uv]
dev-dependencies = []

